requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0
tqdm>=4.65.0
moviepy>=1.0.3
openai>=1.0.0
azure-cognitiveservices-speech>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0 